In this commit: 1. Added a new graph searcher implementation which knows to persist the graph into a file 2. Added 2 files - serialized_edges.pkl and serialized_nodes.pkl which are the output of the persist_graph_searcher.py 3. added GraphLoader class which given filenames, can create a nx.DiGraph instance 4. created benchmark_runner.py which load a graph from the files, and execute the astar_searcher.
TODO: inject heuristic function in benchmark_runner.py. create a file of source and target.

General note:
The goal of the project is to improve the graph search component. In order to do it, in this commit I have made the infrastructure which will allow to isolate the searcher from the whole simulator flow and run it separately. This will allow us to run tests and create benchmarks on different measurements which will prove whether we have improve the algorithm on the new implementation